Module Debug_protocol.Breakpoint

type t = {

id : int option;
An optional identifier for the breakpoint. It is needed if breakpoint events are used to update or remove breakpoints.

verified : bool;
If true breakpoint could be set (but not necessarily at the desired location).

message : string option;
An optional message about the state of the breakpoint. This is shown to the user and can be used to explain why a breakpoint could not be verified.

source : Source.t option;
The source where the breakpoint is located.

line : int option;
The start line of the actual range covered by the breakpoint.

column : int option;
An optional start column of the actual range covered by the breakpoint.

end_line : int option;
An optional end line of the actual range covered by the breakpoint.

end_column : int option;
An optional end column of the actual range covered by the breakpoint. If no end line is given, then the end column is assumed to be in the start line.

instruction_reference : string option;
An optional memory reference to where the breakpoint is set.

offset : int option;
An optional offset from the instruction reference. This can be negative.

}

Information about a Breakpoint created in setBreakpoints, setFunctionBreakpoints, setInstructionBreakpoints, or setDataBreakpoints.

val make : ?⁠id:int option -> verified:bool -> ?⁠message:string option -> ?⁠source:Source.t option -> ?⁠line:int option -> ?⁠column:int option -> ?⁠end_line:int option -> ?⁠end_column:int option -> ?⁠instruction_reference:string option -> ?⁠offset:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or