Module Debug_protocol.Breakpoint_location

type t = {

line : int;
Start line of breakpoint location.

column : int option;
Optional start column of breakpoint location.

end_line : int option;
Optional end line of breakpoint location if the location covers a range.

end_column : int option;
Optional end column of breakpoint location if the location covers a range.

}

Properties of a breakpoint location returned from the 'breakpointLocations' request.

val make : line:int -> ?⁠column:int option -> ?⁠end_line:int option -> ?⁠end_column:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or