Module Breakpoint_locations_command.Arguments

type t = {

source : Source.t;
The source location of the breakpoints; either 'source.path' or 'source.reference' must be specified.

line : int;
Start line of range to search possible breakpoint locations in. If only the line is specified, the request returns all possible locations in that line.

column : int option;
Optional start column of range to search possible breakpoint locations in. If no start column is given, the first column in the start line is assumed.

end_line : int option;
Optional end line of range to search possible breakpoint locations in. If no end line is given, then the end line is assumed to be the start line.

end_column : int option;
Optional end column of range to search possible breakpoint locations in. If no end column is given, then it is assumed to be in the last column of the end line.

}

Arguments for 'breakpointLocations' request.

val make : source:Source.t -> line:int -> ?⁠column:int option -> ?⁠end_line:int option -> ?⁠end_column:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or