Module Debug_protocol.Column_descriptor

module Type : sig ... end

type t = {

attribute_name : string;
Name of the attribute rendered in this column.

label : string;
Header UI label of column.

format : string option;
Format to use for the rendered values in this column. TBD how the format strings looks like.

type_ : Type.t option;
Datatype of values in this column. Defaults to 'string' if not specified.

width : int option;
Width of this column in characters (hint only).

}

A ColumnDescriptor specifies what module attribute to show in a column of the ModulesView, how to format it, and what the column's label should be. It is only used if the underlying UI actually supports this level of customization.

val make : attribute_name:string -> label:string -> ?⁠format:string option -> ?⁠type_:Type.t option -> ?⁠width:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or