Module Debug_protocol.Configuration_done_command

This optional request indicates that the client has finished initialization of the debug adapter. So it is the last request in the sequence of configuration requests (which was started by the 'initialized' event). Clients should only call this request if the capability 'supportsConfigurationDoneRequest' is true.

val type_ : string

module Arguments : sig ... end

module Result : sig ... end