Module Debug_protocol.Data_breakpoint

type t = {

data_id : string;
An id representing the data. This id is returned from the dataBreakpointInfo request.

access_type : Data_breakpoint_access_type.t option;
The access type of the data.

condition : string option;
An optional expression for conditional breakpoints.

hit_condition : string option;
An optional expression that controls how many hits of the breakpoint are ignored. The backend is expected to interpret the expression as needed.

}

Properties of a data breakpoint passed to the setDataBreakpoints request.

val make : data_id:string -> ?⁠access_type:Data_breakpoint_access_type.t option -> ?⁠condition:string option -> ?⁠hit_condition:string option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or