Module Data_breakpoint_info_command.Result

type t = {

data_id : string option;
An identifier for the data on which a data breakpoint can be registered with the setDataBreakpoints request or null if no data breakpoint is available.

description : string;
UI string that describes on what data the breakpoint is set on or why a data breakpoint is not available.

access_types : Data_breakpoint_access_type.t list option;
Optional attribute listing the available access types for a potential data breakpoint. A UI frontend could surface this information.

can_persist : bool option;
Optional attribute indicating that a potential data breakpoint could be persisted across sessions.

}

val make : ?⁠data_id:string -> description:string -> ?⁠access_types:Data_breakpoint_access_type.t list option -> ?⁠can_persist:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or