Module Disassemble_command.Arguments

type t = {

memory_reference : string;
Memory reference to the base location containing the instructions to disassemble.

offset : int option;
Optional offset (in bytes) to be applied to the reference location before disassembling. Can be negative.

instruction_offset : int option;
Optional offset (in instructions) to be applied after the byte offset (if any) before disassembling. Can be negative.

instruction_count : int;
Number of instructions to disassemble starting at the specified location and offset. An adapter must return exactly this number of instructions - any unavailable instructions should be replaced with an implementation-defined 'invalid instruction' value.

resolve_symbols : bool option;
If true, the adapter should attempt to resolve memory addresses and other values to symbolic names.

}

Arguments for 'disassemble' request.

val make : memory_reference:string -> ?⁠offset:int option -> ?⁠instruction_offset:int option -> instruction_count:int -> ?⁠resolve_symbols:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or