Module Debug_protocol.Disassembled_instruction

type t = {

address : string;
The address of the instruction. Treated as a hex value if prefixed with '0x', or as a decimal value otherwise.

instruction_bytes : string option;
Optional raw bytes representing the instruction and its operands, in an implementation-defined format.

instruction : string;
Text representing the instruction and its operands, in an implementation-defined format.

symbol : string option;
Name of the symbol that corresponds with the location of this instruction, if any.

location : Source.t option;
Source location that corresponds to this instruction, if any. Should always be set (if available) on the first instruction returned, but can be omitted afterwards if this instruction maps to the same source file as the previous instruction.

line : int option;
The line within the source location that corresponds to this instruction, if any.

column : int option;
The column within the line that corresponds to this instruction, if any.

end_line : int option;
The end line of the range that corresponds to this instruction, if any.

end_column : int option;
The end column of the range that corresponds to this instruction, if any.

}

Represents a single disassembled instruction.

val make : address:string -> ?⁠instruction_bytes:string option -> instruction:string -> ?⁠symbol:string option -> ?⁠location:Source.t option -> ?⁠line:int option -> ?⁠column:int option -> ?⁠end_line:int option -> ?⁠end_column:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or