Module Disconnect_command.Arguments

type t = {

restart : bool option;
A value of true indicates that this 'disconnect' request is part of a restart sequence.

terminate_debuggee : bool option;
Indicates whether the debuggee should be terminated when the debugger is disconnected. If unspecified, the debug adapter is free to do whatever it thinks is best. The attribute is only honored by a debug adapter if the capability 'supportTerminateDebuggee' is true.

}

Arguments for 'disconnect' request.

val make : ?⁠restart:bool option -> ?⁠terminate_debuggee:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or