Module Evaluate_command.Arguments

module Context : sig ... end

type t = {

expression : string;
The expression to evaluate.

frame_id : int option;
Evaluate the expression in the scope of this stack frame. If not specified, the expression is evaluated in the global scope.

context : Context.t option;
The context in which the evaluate request is run.

format : Value_format.t option;
Specifies details on how to format the Evaluate result. The attribute is only honored by a debug adapter if the capability 'supportsValueFormattingOptions' is true.

}

Arguments for 'evaluate' request.

val make : expression:string -> ?⁠frame_id:int option -> ?⁠context:Context.t option -> ?⁠format:Value_format.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or