Module Evaluate_command.Result

type t = {

result : string;
The result of the evaluate request.

type_ : string option;
The optional type of the evaluate result. This attribute should only be returned by a debug adapter if the client has passed the value true for the 'supportsVariableType' capability of the 'initialize' request.

presentation_hint : Variable_presentation_hint.t option;
Properties of a evaluate result that can be used to determine how to render the result in the UI.

variables_reference : int;
If variablesReference is > 0, the evaluate result is structured and its children can be retrieved by passing variablesReference to the VariablesRequest. The value should be less than or equal to 2147483647 (2^31-1).

named_variables : int option;
The number of named child variables. The client can use this optional information to present the variables in a paged UI and fetch them in chunks. The value should be less than or equal to 2147483647 (2^31-1).

indexed_variables : int option;
The number of indexed child variables. The client can use this optional information to present the variables in a paged UI and fetch them in chunks. The value should be less than or equal to 2147483647 (2^31-1).

memory_reference : string option;
Optional memory reference to a location appropriate for this result. For pointer type eval results, this is generally a reference to the memory address contained in the pointer. This attribute should be returned by a debug adapter if the client has passed the value true for the 'supportsMemoryReferences' capability of the 'initialize' request.

}

val make : result:string -> ?⁠type_:string option -> ?⁠presentation_hint:Variable_presentation_hint.t option -> variables_reference:int -> ?⁠named_variables:int option -> ?⁠indexed_variables:int option -> ?⁠memory_reference:string option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or