Module Debug_protocol.Exception_breakpoints_filter

type t = {

filter : string;
The internal ID of the filter option. This value is passed to the 'setExceptionBreakpoints' request.

label : string;
The name of the filter option. This will be shown in the UI.

default : bool option;
Initial value of the filter option. If not specified a value 'false' is assumed.

supports_condition : bool option;
Controls whether a condition can be specified for this filter option. If false or missing, a condition can not be set.

}

An ExceptionBreakpointsFilter is shown in the UI as an filter option for configuring how exceptions are dealt with.

val make : filter:string -> label:string -> ?⁠default:bool option -> ?⁠supports_condition:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or