Module Debug_protocol.Exception_details

type t = {

message : string option;
Message contained in the exception.

type_name : string option;
Short type name of the exception object.

full_type_name : string option;
Fully-qualified type name of the exception object.

evaluate_name : string option;
Optional expression that can be evaluated in the current scope to obtain the exception object.

stack_trace : string option;
Stack trace at the time the exception was thrown.

inner_exception : t option;
Details of the exception contained by this exception, if any.

}

Detailed information about an exception that has occurred.

val make : ?⁠message:string option -> ?⁠type_name:string option -> ?⁠full_type_name:string option -> ?⁠evaluate_name:string option -> ?⁠stack_trace:string option -> ?⁠inner_exception:t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or