Module Debug_protocol.Exception_info_command

Retrieves the details of the exception that caused this event to be raised. Clients should only call this request if the capability 'supportsExceptionInfoRequest' is true.

val type_ : string

module Arguments : sig ... end

module Result : sig ... end