Module Exception_info_command.Result

type t = {

exception_id : string;
ID of the exception that was thrown.

description : string option;
Descriptive text for the exception provided by the debug adapter.

break_mode : Exception_break_mode.t;
Mode that caused the exception notification to be raised.

details : Exception_details.t option;
Detailed information about the exception.

}

val make : exception_id:string -> ?⁠description:string option -> break_mode:Exception_break_mode.t -> ?⁠details:Exception_details.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or