Module Debug_protocol.Exception_options

type t = {

path : Exception_path_segment.t list option;
A path that selects a single or multiple exceptions in a tree. If 'path' is missing, the whole tree is selected. By convention the first segment of the path is a category that is used to group exceptions in the UI.

break_mode : Exception_break_mode.t;
Condition when a thrown exception should result in a break.

}

An ExceptionOptions assigns configuration options to a set of exceptions.

val make : ?⁠path:Exception_path_segment.t list option -> break_mode:Exception_break_mode.t -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or