Module Debug_protocol.Exception_path_segment

type t = {

negate : bool option;
If false or missing this segment matches the names provided, otherwise it matches anything except the names provided.

names : string;
Depending on the value of 'negate' the names that should match or not match.

}

An ExceptionPathSegment represents a segment in a path that is used to match leafs or nodes in a tree of exceptions. If a segment consists of more than one name, it matches the names provided if 'negate' is false or missing or it matches anything except the names provided if 'negate' is true.

val make : ?⁠negate:bool option -> names:string -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or