Module Debug_protocol.Goto_target

type t = {

id : int;
Unique identifier for a goto target. This is used in the goto request.

label : string;
The name of the goto target (shown in the UI).

line : int;
The line of the goto target.

column : int option;
An optional column of the goto target.

end_line : int option;
An optional end line of the range covered by the goto target.

end_column : int option;
An optional end column of the range covered by the goto target.

instruction_pointer_reference : string option;
Optional memory reference for the instruction pointer value represented by this target.

}

A GotoTarget describes a code location that can be used as a target in the 'goto' request. The possible goto targets can be determined via the 'gotoTargets' request.

val make : id:int -> label:string -> line:int -> ?⁠column:int option -> ?⁠end_line:int option -> ?⁠end_column:int option -> ?⁠instruction_pointer_reference:string option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or