Module Goto_targets_command.Arguments

type t = {

source : Source.t;
The source location for which the goto targets are determined.

line : int;
The line location for which the goto targets are determined.

column : int option;
An optional column location for which the goto targets are determined.

}

Arguments for 'gotoTargets' request.

val make : source:Source.t -> line:int -> ?⁠column:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or