Module Initialize_command.Result

type t = {

supports_configuration_done_request : bool option;
The debug adapter supports the 'configurationDone' request.

supports_function_breakpoints : bool option;
The debug adapter supports function breakpoints.

supports_conditional_breakpoints : bool option;
The debug adapter supports conditional breakpoints.

supports_hit_conditional_breakpoints : bool option;
The debug adapter supports breakpoints that break execution after a specified number of hits.

supports_evaluate_for_hovers : bool option;
The debug adapter supports a (side effect free) evaluate request for data hovers.

exception_breakpoint_filters : Exception_breakpoints_filter.t option;
Available exception filter options for the 'setExceptionBreakpoints' request.

supports_step_back : bool option;
The debug adapter supports stepping back via the 'stepBack' and 'reverseContinue' requests.

supports_set_variable : bool option;
The debug adapter supports setting a variable to a value.

supports_restart_frame : bool option;
The debug adapter supports restarting a frame.

supports_goto_targets_request : bool option;
The debug adapter supports the 'gotoTargets' request.

supports_step_in_targets_request : bool option;
The debug adapter supports the 'stepInTargets' request.

supports_completions_request : bool option;
The debug adapter supports the 'completions' request.

completion_trigger_characters : string option;
The set of characters that should trigger completion in a REPL. If not specified, the UI should assume the '.' character.

supports_modules_request : bool option;
The debug adapter supports the 'modules' request.

additional_module_columns : Column_descriptor.t option;
The set of additional module information exposed by the debug adapter.

supported_checksum_algorithms : Checksum_algorithm.t option;
Checksum algorithms supported by the debug adapter.

supports_restart_request : bool option;
The debug adapter supports the 'restart' request. In this case a client should not implement 'restart' by terminating and relaunching the adapter but by calling the RestartRequest.

supports_exception_options : bool option;
The debug adapter supports 'exceptionOptions' on the setExceptionBreakpoints request.

supports_value_formatting_options : bool option;
The debug adapter supports a 'format' attribute on the stackTraceRequest, variablesRequest, and evaluateRequest.

supports_exception_info_request : bool option;
The debug adapter supports the 'exceptionInfo' request.

support_terminate_debuggee : bool option;
The debug adapter supports the 'terminateDebuggee' attribute on the 'disconnect' request.

supports_delayed_stack_trace_loading : bool option;
The debug adapter supports the delayed loading of parts of the stack, which requires that both the 'startFrame' and 'levels' arguments and the 'totalFrames' result of the 'StackTrace' request are supported.

supports_loaded_sources_request : bool option;
The debug adapter supports the 'loadedSources' request.

supports_log_points : bool option;
The debug adapter supports logpoints by interpreting the 'logMessage' attribute of the SourceBreakpoint.

supports_terminate_threads_request : bool option;
The debug adapter supports the 'terminateThreads' request.

supports_set_expression : bool option;
The debug adapter supports the 'setExpression' request.

supports_terminate_request : bool option;
The debug adapter supports the 'terminate' request.

supports_data_breakpoints : bool option;
The debug adapter supports data breakpoints.

supports_read_memory_request : bool option;
The debug adapter supports the 'readMemory' request.

supports_disassemble_request : bool option;
The debug adapter supports the 'disassemble' request.

supports_cancel_request : bool option;
The debug adapter supports the 'cancel' request.

supports_breakpoint_locations_request : bool option;
The debug adapter supports the 'breakpointLocations' request.

supports_clipboard_context : bool option;
The debug adapter supports the 'clipboard' context value in the 'evaluate' request.

supports_stepping_granularity : bool option;
The debug adapter supports stepping granularities (argument 'granularity') for the stepping requests.

supports_instruction_breakpoints : bool option;
The debug adapter supports adding breakpoints based on instruction references.

supports_exception_filter_options : bool option;
The debug adapter supports 'filterOptions' as an argument on the 'setExceptionBreakpoints' request.

}

Information about the capabilities of a debug adapter.

val make : ?⁠supports_configuration_done_request:bool option -> ?⁠supports_function_breakpoints:bool option -> ?⁠supports_conditional_breakpoints:bool option -> ?⁠supports_hit_conditional_breakpoints:bool option -> ?⁠supports_evaluate_for_hovers:bool option -> ?⁠exception_breakpoint_filters:Exception_breakpoints_filter.t option -> ?⁠supports_step_back:bool option -> ?⁠supports_set_variable:bool option -> ?⁠supports_restart_frame:bool option -> ?⁠supports_goto_targets_request:bool option -> ?⁠supports_step_in_targets_request:bool option -> ?⁠supports_completions_request:bool option -> ?⁠completion_trigger_characters:string option -> ?⁠supports_modules_request:bool option -> ?⁠additional_module_columns:Column_descriptor.t option -> ?⁠supported_checksum_algorithms:Checksum_algorithm.t option -> ?⁠supports_restart_request:bool option -> ?⁠supports_exception_options:bool option -> ?⁠supports_value_formatting_options:bool option -> ?⁠supports_exception_info_request:bool option -> ?⁠support_terminate_debuggee:bool option -> ?⁠supports_delayed_stack_trace_loading:bool option -> ?⁠supports_loaded_sources_request:bool option -> ?⁠supports_log_points:bool option -> ?⁠supports_terminate_threads_request:bool option -> ?⁠supports_set_expression:bool option -> ?⁠supports_terminate_request:bool option -> ?⁠supports_data_breakpoints:bool option -> ?⁠supports_read_memory_request:bool option -> ?⁠supports_disassemble_request:bool option -> ?⁠supports_cancel_request:bool option -> ?⁠supports_breakpoint_locations_request:bool option -> ?⁠supports_clipboard_context:bool option -> ?⁠supports_stepping_granularity:bool option -> ?⁠supports_instruction_breakpoints:bool option -> ?⁠supports_exception_filter_options:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or