Module Invalidated_event.Payload

type t = {

areas : Invalidated_areas.t list option;
Optional set of logical areas that got invalidated. This property has a hint characteristic: a client can only be expected to make a 'best effort' in honouring the areas but there are no guarantees. If this property is missing, empty, or if values are not understand the client should assume a single value 'all'.

thread_id : int option;
If specified, the client only needs to refetch data related to this thread.

stack_frame_id : int option;
If specified, the client only needs to refetch data related to this stack frame (and the 'threadId' is ignored).

}

val make : ?⁠areas:Invalidated_areas.t list option -> ?⁠thread_id:int option -> ?⁠stack_frame_id:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or