Module Debug_protocol.Message

module Variables : sig ... end

type t = {

id : int;
Unique identifier for the message.

format : string;
A format string for the message. Embedded variables have the form '{name}'. If variable name starts with an underscore character, the variable does not contain user data (PII) and can be safely used for telemetry purposes.

variables : Variables.t option;
An object used as a dictionary for looking up the variables in the format string.

send_telemetry : bool option;
If true send to telemetry.

show_user : bool option;
If true show user.

url : string option;
An optional url where additional information about this message can be found.

url_label : string option;
An optional label that is presented to the user as the UI for opening the url.

}

A structured message object. Used to return errors from requests.

val make : id:int -> format:string -> ?⁠variables:Variables.t option -> ?⁠send_telemetry:bool option -> ?⁠show_user:bool option -> ?⁠url:string option -> ?⁠url_label:string option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or