Module Debug_protocol.Modules_view_descriptor

type t = {

columns : Column_descriptor.t;

}

The ModulesViewDescriptor is the container for all declarative configuration options of a ModuleView. For now it only specifies the columns to be shown in the modules view.

val make : columns:Column_descriptor.t -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or