Module Debug_protocol.Next_command

The request starts the debuggee to run again for one step. The debug adapter first sends the response and then a 'stopped' event (with reason 'step') after the step has completed.

val type_ : string

module Arguments : sig ... end

module Result : sig ... end