Module Output_event.Payload

module Category : sig ... end

module Group : sig ... end

type t = {

category : Category.t option;
The output category. If not specified, 'console' is assumed.

output : string;
The output to report.

group : Group.t option;
Support for keeping an output log organized by grouping related messages.

variables_reference : int option;
If an attribute 'variablesReference' exists and its value is > 0, the output contains objects which can be retrieved by passing 'variablesReference' to the 'variables' request. The value should be less than or equal to 2147483647 (2^31-1).

source : Source.t option;
An optional source location where the output was produced.

line : int option;
An optional source location line where the output was produced.

column : int option;
An optional source location column where the output was produced.

data : Any.t option;
Optional data to report. For the 'telemetry' category the data will be sent to telemetry, for the other categories the data is shown in JSON format.

}

val make : ?⁠category:Category.t option -> output:string -> ?⁠group:Group.t option -> ?⁠variables_reference:int option -> ?⁠source:Source.t option -> ?⁠line:int option -> ?⁠column:int option -> ?⁠data:Any.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or