Module Process_event.Payload

module Start_method : sig ... end

type t = {

name : string;
The logical name of the process. This is usually the full path to process's executable file. Example: /home/example/myproj/program.js.

system_process_id : int option;
The system process id of the debugged process. This property will be missing for non-system processes.

is_local_process : bool option;
If true, the process is running on the same computer as the debug adapter.

start_method : Start_method.t option;
Describes how the debug engine started debugging this process.

pointer_size : int option;
The size of a pointer or address for this process, in bits. This value may be used by clients when formatting addresses for display.

}

val make : name:string -> ?⁠system_process_id:int option -> ?⁠is_local_process:bool option -> ?⁠start_method:Start_method.t option -> ?⁠pointer_size:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or