Module Progress_start_event.Payload

type t = {

progress_id : string;
An ID that must be used in subsequent 'progressUpdate' and 'progressEnd' events to make them refer to the same progress reporting. IDs must be unique within a debug session.

title : string;
Mandatory (short) title of the progress reporting. Shown in the UI to describe the long running operation.

request_id : int option;
The request ID that this progress report is related to. If specified a debug adapter is expected to emit progress events for the long running request until the request has been either completed or cancelled. If the request ID is omitted, the progress report is assumed to be related to some general activity of the debug adapter.

cancellable : bool option;
If true, the request that reports progress may be canceled with a 'cancel' request. So this property basically controls whether the client should use UX that supports cancellation. Clients that don't support cancellation are allowed to ignore the setting.

message : string option;
Optional, more detailed progress message.

percentage : float option;
Optional progress percentage to display (value range: 0 to 100). If omitted no percentage will be shown.

}

val make : progress_id:string -> title:string -> ?⁠request_id:int option -> ?⁠cancellable:bool option -> ?⁠message:string option -> ?⁠percentage:float option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or