Module Read_memory_command.Arguments

type t = {

memory_reference : string;
Memory reference to the base location from which data should be read.

offset : int option;
Optional offset (in bytes) to be applied to the reference location before reading data. Can be negative.

count : int;
Number of bytes to read at the specified location and offset.

}

Arguments for 'readMemory' request.

val make : memory_reference:string -> ?⁠offset:int option -> count:int -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or