Module Read_memory_command.Result

type t = {

address : string;
The address of the first byte of data returned. Treated as a hex value if prefixed with '0x', or as a decimal value otherwise.

unreadable_bytes : int option;
The number of unreadable bytes encountered after the last successfully read byte. This can be used to determine the number of bytes that must be skipped before a subsequent 'readMemory' request will succeed.

data : string option;
The bytes read from memory, encoded using base64.

}

val make : address:string -> ?⁠unreadable_bytes:int option -> ?⁠data:string option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or