Module Debug_protocol.Response

module Type : sig ... end

module Message : sig ... end

type t = {

seq : int;
Sequence number (also known as message ID). For protocol messages of type 'request' this ID can be used to cancel the request.

type_ : Type.t;

request_seq : int;
Sequence number of the corresponding request.

success : bool;
Outcome of the request. If true, the request was successful and the 'body' attribute may contain the result of the request. If the value is false, the attribute 'message' contains the error in short form and the 'body' may contain additional information (see 'ErrorResponse.body.error').

command : string;
The command requested.

message : Message.t option;
Contains the raw error in short form if 'success' is false. This raw error might be interpreted by the frontend and is not shown in the UI. Some predefined values exist.

body : Yojson.Safe.t;
Contains request result if success is true and optional error details if success is false.

}

val make : seq:int -> type_:Type.t -> request_seq:int -> success:bool -> command:string -> ?⁠message:Message.t option -> ?⁠body:Yojson.Safe.t -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or