Module Response.Message

type t =

| Cancelled

| Custom of string

Contains the raw error in short form if 'success' is false. This raw error might be interpreted by the frontend and is not shown in the UI. Some predefined values exist.

include JSONABLE with type t := t

type t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val to_yojson : t -> Yojson.Safe.t