Module Debug_protocol.Restart_command

Restarts a debug session. Clients should only call this request if the capability 'supportsRestartRequest' is true. If the capability is missing or has the value false, a typical client will emulate 'restart' by terminating the debug adapter first and then launching it anew.

val type_ : string

module Arguments : sig ... end

module Result : sig ... end