Module Run_in_terminal_command.Arguments

module Kind : sig ... end

module Env : sig ... end

type t = {

kind : Kind.t option;
What kind of terminal to launch.

title : string option;
Optional title of the terminal.

cwd : string;
Working directory for the command. For non-empty, valid paths this typically results in execution of a change directory command.

args : string;
List of arguments. The first argument is the command to run.

env : Env.t option;
Environment key-value pairs that are added to or removed from the default environment.

}

Arguments for 'runInTerminal' request.

val make : ?⁠kind:Kind.t option -> ?⁠title:string option -> cwd:string -> args:string -> ?⁠env:Env.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or