Module Run_in_terminal_command.Result

type t = {

process_id : int option;
The process ID. The value should be less than or equal to 2147483647 (2^31-1).

shell_process_id : int option;
The process ID of the terminal shell. The value should be less than or equal to 2147483647 (2^31-1).

}

val make : ?⁠process_id:int option -> ?⁠shell_process_id:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or