Module Debug_protocol.Set_breakpoints_command

Sets multiple breakpoints for a single source and clears all previous breakpoints in that source. To clear all breakpoint for a source, specify an empty array. When a breakpoint is hit, a 'stopped' event (with reason 'breakpoint') is generated.

val type_ : string

module Arguments : sig ... end

module Result : sig ... end