Module Set_breakpoints_command.Arguments

type t = {

source : Source.t;
The source location of the breakpoints; either 'source.path' or 'source.reference' must be specified.

breakpoints : Source_breakpoint.t list option;
The code locations of the breakpoints.

lines : int list option;
Deprecated: The code locations of the breakpoints.

source_modified : bool option;
A value of true indicates that the underlying source has been modified which results in new breakpoint locations.

}

Arguments for 'setBreakpoints' request.

val make : source:Source.t -> ?⁠breakpoints:Source_breakpoint.t list option -> ?⁠lines:int list option -> ?⁠source_modified:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or