Module Set_expression_command.Arguments

type t = {

expression : string;
The l-value expression to assign to.

value : string;
The value expression to assign to the l-value expression.

frame_id : int option;
Evaluate the expressions in the scope of this stack frame. If not specified, the expressions are evaluated in the global scope.

format : Value_format.t option;
Specifies how the resulting value should be formatted.

}

Arguments for 'setExpression' request.

val make : expression:string -> value:string -> ?⁠frame_id:int option -> ?⁠format:Value_format.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or