Module Set_variable_command.Result

type t = {

value : string;
The new value of the variable.

type_ : string option;
The type of the new value. Typically shown in the UI when hovering over the value.

variables_reference : int option;
If variablesReference is > 0, the new value is structured and its children can be retrieved by passing variablesReference to the VariablesRequest. The value should be less than or equal to 2147483647 (2^31-1).

named_variables : int option;
The number of named child variables. The client can use this optional information to present the variables in a paged UI and fetch them in chunks. The value should be less than or equal to 2147483647 (2^31-1).

indexed_variables : int option;
The number of indexed child variables. The client can use this optional information to present the variables in a paged UI and fetch them in chunks. The value should be less than or equal to 2147483647 (2^31-1).

}

val make : value:string -> ?⁠type_:string option -> ?⁠variables_reference:int option -> ?⁠named_variables:int option -> ?⁠indexed_variables:int option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or