Module Debug_protocol.Source

module Presentation_hint : sig ... end

type t = {

name : string option;
The short name of the source. Every source returned from the debug adapter has a name. When sending a source to the debug adapter this name is optional.

path : string option;
The path of the source to be shown in the UI. It is only used to locate and load the content of the source if no sourceReference is specified (or its value is 0).

source_reference : int option;
If sourceReference > 0 the contents of the source must be retrieved through the SourceRequest (even if a path is specified). A sourceReference is only valid for a session, so it must not be used to persist a source. The value should be less than or equal to 2147483647 (2^31-1).

presentation_hint : Presentation_hint.t option;
An optional hint for how to present the source in the UI. A value of 'deemphasize' can be used to indicate that the source is not available or that it is skipped on stepping.

origin : string option;
The (optional) origin of this source: possible values 'internal module', 'inlined content from source map', etc.

sources : t option;
An optional list of sources that are related to this source. These may be the source that generated this source.

adapter_data : Any.t option;
Optional data that a debug adapter might want to loop through the client. The client should leave the data intact and persist it across sessions. The client should not interpret the data.

checksums : Checksum.t option;
The checksums associated with this file.

}

A Source is a descriptor for source code. It is returned from the debug adapter as part of a StackFrame and it is used by clients when specifying breakpoints.

val make : ?⁠name:string option -> ?⁠path:string option -> ?⁠source_reference:int option -> ?⁠presentation_hint:Presentation_hint.t option -> ?⁠origin:string option -> ?⁠sources:t option -> ?⁠adapter_data:Any.t option -> ?⁠checksums:Checksum.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or