Module Source.Presentation_hint

type t =

| Normal

| Emphasize

| Deemphasize

An optional hint for how to present the source in the UI. A value of 'deemphasize' can be used to indicate that the source is not available or that it is skipped on stepping.

include JSONABLE with type t := t

type t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val to_yojson : t -> Yojson.Safe.t