Module Debug_protocol.Stack_frame

module Presentation_hint : sig ... end

type t = {

id : int;
An identifier for the stack frame. It must be unique across all threads. This id can be used to retrieve the scopes of the frame with the 'scopesRequest' or to restart the execution of a stackframe.

name : string;
The name of the stack frame, typically a method name.

source : Source.t option;
The optional source of the frame.

line : int;
The line within the file of the frame. If source is null or doesn't exist, line is 0 and must be ignored.

column : int;
The column within the line. If source is null or doesn't exist, column is 0 and must be ignored.

end_line : int option;
An optional end line of the range covered by the stack frame.

end_column : int option;
An optional end column of the range covered by the stack frame.

instruction_pointer_reference : string option;
Optional memory reference for the current instruction pointer in this frame.

module_id : Int_or_string.t option;
The module associated with this frame, if any.

presentation_hint : Presentation_hint.t option;
An optional hint for how to present this frame in the UI. A value of 'label' can be used to indicate that the frame is an artificial frame that is used as a visual label or separator. A value of 'subtle' can be used to change the appearance of a frame in a 'subtle' way.

}

A Stackframe contains the source location.

val make : id:int -> name:string -> ?⁠source:Source.t option -> line:int -> column:int -> ?⁠end_line:int option -> ?⁠end_column:int option -> ?⁠instruction_pointer_reference:string option -> ?⁠module_id:Int_or_string.t option -> ?⁠presentation_hint:Presentation_hint.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or