Module Debug_protocol.Stack_frame_format

type t = {

hex : bool option;
Display the value in hex.

parameters : bool option;
Displays parameters for the stack frame.

parameter_types : bool option;
Displays the types of parameters for the stack frame.

parameter_names : bool option;
Displays the names of parameters for the stack frame.

parameter_values : bool option;
Displays the values of parameters for the stack frame.

line : bool option;
Displays the line number of the stack frame.

module_ : bool option;
Displays the module of the stack frame.

include_all : bool option;
Includes all stack frames, including those the debug adapter might otherwise hide.

}

val make : ?⁠hex:bool option -> ?⁠parameters:bool option -> ?⁠parameter_types:bool option -> ?⁠parameter_names:bool option -> ?⁠parameter_values:bool option -> ?⁠line:bool option -> ?⁠module_:bool option -> ?⁠include_all:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or