Module Stack_trace_command.Arguments

type t = {

thread_id : int;
Retrieve the stacktrace for this thread.

start_frame : int option;
The index of the first frame to return; if omitted frames start at 0.

levels : int option;
The maximum number of frames to return. If levels is not specified or 0, all frames are returned.

format : Stack_frame_format.t option;
Specifies details on how to format the stack frames. The attribute is only honored by a debug adapter if the capability 'supportsValueFormattingOptions' is true.

}

Arguments for 'stackTrace' request.

val make : thread_id:int -> ?⁠start_frame:int option -> ?⁠levels:int option -> ?⁠format:Stack_frame_format.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or