Module Debug_protocol.Variable_presentation_hint

module Kind : sig ... end

module Attributes : sig ... end

module Visibility : sig ... end

type t = {

kind : Kind.t option;
The kind of variable. Before introducing additional values, try to use the listed values.

attributes : Attributes.t list option;
Set of attributes represented as an array of strings. Before introducing additional values, try to use the listed values.

visibility : Visibility.t option;
Visibility of variable. Before introducing additional values, try to use the listed values.

}

Optional properties of a variable that can be used to determine how to render the variable in the UI.

val make : ?⁠kind:Kind.t option -> ?⁠attributes:Attributes.t list option -> ?⁠visibility:Visibility.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or