Module Variables_command.Arguments

module Filter : sig ... end

type t = {

variables_reference : int;
The Variable reference.

filter : Filter.t option;
Optional filter to limit the child variables to either named or indexed. If omitted, both types are fetched.

start : int option;
The index of the first variable to return; if omitted children start at 0.

count : int option;
The number of variables to return. If count is missing or 0, all variables are returned.

format : Value_format.t option;
Specifies details on how to format the Variable values. The attribute is only honored by a debug adapter if the capability 'supportsValueFormattingOptions' is true.

}

Arguments for 'variables' request.

val make : variables_reference:int -> ?⁠filter:Filter.t option -> ?⁠start:int option -> ?⁠count:int option -> ?⁠format:Value_format.t option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or