Module Stopped_event.Payload

module Reason : sig ... end

type t = {

reason : Reason.t;
The reason for the event. For backward compatibility this string is shown in the UI if the 'description' attribute is missing (but it must not be translated).

description : string option;
The full reason for the event, e.g. 'Paused on exception'. This string is shown in the UI as is and must be translated.

thread_id : int option;
The thread which was stopped.

preserve_focus_hint : bool option;
A value of true hints to the frontend that this event should not change the focus.

text : string option;
Additional information. E.g. if reason is 'exception', text contains the exception name. This string is shown in the UI.

all_threads_stopped : bool option;
If 'allThreadsStopped' is true, a debug adapter can announce that all threads have stopped.

The client should use this information to enable that all threads can be expanded to access their stacktraces.

If the attribute is missing or false, only the thread with the given threadId can be expanded.

}

val make : reason:Reason.t -> ?⁠description:string option -> ?⁠thread_id:int option -> ?⁠preserve_focus_hint:bool option -> ?⁠text:string option -> ?⁠all_threads_stopped:bool option -> unit -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or